requirements.txt
bcrypt==3.2.0
cffi==1.14.5
click==7.1.2
dnspython==2.1.0
fastapi==0.63.0
h11==0.12.0
passlib==1.7.4
pycparser==2.20
pydantic==1.8.1
PyJWT==2.0.1
pymongo==3.11.3
python-dotenv==0.15.0
six==1.15.0
starlette==0.13.6
typing-extensions==3.7.4.3
uvicorn==0.13.4